Update Message.cs serialization for better in editor list layout (#94)
What I changed:
Updated Message.cs so that the first parameter is the name variable.

Reason:
When creating a list of messages using "List<Message>() messages = new
List<Message>();" it creates a list of message objects in the inspector
but each message on the list shows as 'element 0', 'element 1', 'element
2' etc., but if the first parameter is a string called 'name' the
messages on the list will be named after the parameter name.

This is useful because now we can have all the messages on our list of
messages in the inspector collapsed and still see what they are called
which means we don't need to have every item on the list expanded to see
what is what.

It looks nicer, it's more convenient and will allow for a slightly
faster workflow when using lists of messages in the inspector. Also,
just why not? The variable was already there and so it only makes sense.

(PS I am using the OpenUPM version so I can't change the script myself
even in my own project so you will need to test this first just in case
of human error)